Completed A3 Java project
The purpose of this software design exercise is to implement a specication for software modules to support the process of CEAB (Canadian Engineering Board) accreditation.
The specication is given in the le A3P1 Spec.pdf.   All of your code should be written in Java. At least CourseT.java should be documented using doxygen. Your code should follow the given specification exactly. In particular, you should not add public methods or procedures that are not specified and you should not change the number or order of parameters for methods or procedures. 
The intention is that your code will be run at the command line (terminal).